move autoplot app into plotr package
